requirements.txt
Flask-JWT-Extended
Flask-RESTX
flask-smorest
Flask-SQLAlchemy
passlib
marshmallow
python-dotenv
db